REFACTOR: complex data manips moved into fetcher (mixed gambles stuff)
