+++ nilearn/datasets/func.py
@@ -3,8 +3,9 @@ Downloading NeuroImaging datasets: functional datasets (task + resting-state)
 """
 import warnings
 import os
-import numpy as np
 import re
+import numpy as np
+import nibabel
 from sklearn.datasets.base import Bunch
 
 from .utils import (_get_dataset_dir, _fetch_files, _get_dataset_descr,
@@ -1156,7 +1157,7 @@ def fetch_abide_pcp(data_dir=None, n_subjects=None, pipeline='cpac',
 
 
 def fetch_mixed_gambles(n_subjects=1, data_dir=None, url=None, resume=True,
-                        verbose=0):
+                        return_raw_data=False, verbose=0):
     """Fetch Jimura "mixed gambles" dataset
 
 
@@ -1180,7 +1181,7 @@ def fetch_mixed_gambles(n_subjects=1, data_dir=None, url=None, resume=True,
     verbose: int, optional (default 0)
         Defines the level of verbosity of the output.
 
-    make_Xy: bool, optional (default False)
+    return_raw_data: bool, optional (default False)
         If true, then the data will transformed into and (X, y) pair, suitable
         for machine learning routines. X is a list of n_subjects * 48
         Nifti1Image objects (where 48 is the number of trials),
@@ -1219,7 +1220,45 @@ def fetch_mixed_gambles(n_subjects=1, data_dir=None, url=None, resume=True,
              for j in range(n_subjects)]
     data_dir = _get_dataset_dir('jimura_poldrack_2012_zmaps',
                                 data_dir=data_dir)
-    zmap_fnames = _fetch_files(data_dir, files, resume=resume,
-                               verbose=verbose)
+    zmap_fnames = _fetch_files(data_dir, files, resume=resume, verbose=verbose)
     data = Bunch(zmaps=zmap_fnames)
+    if return_raw_data:
+        X = []
+        y = []
+        mask = []
+        for zmap_fname in data.zmaps:
+            # load subject data
+            img = nibabel.load(zmap_fname)
+            this_X = img.get_data()
+            affine = img.get_affine()
+            finite_mask = np.all(np.isfinite(this_X), axis=-1)
+            this_mask = np.logical_and(np.all(this_X != 0, axis=-1),
+                                       finite_mask)
+            this_y = np.array([np.arange(1, 9)] * 6).ravel()
+
+            # gain levels
+            if len(this_y) != this_X.shape[-1]:
+                raise RuntimeError("%s: Expecting %i volumes, got %i!" % (
+                    zmap_fname, len(this_y), this_X.shape[-1]))
+
+            # standardize subject data
+            this_X -= this_X.mean(axis=-1)[..., np.newaxis]
+            std = this_X.std(axis=-1)
+            std[std == 0] = 1
+            this_X /= std[..., np.newaxis]
+
+            # commit subject data
+            X.append(this_X)
+            y.extend(this_y)
+            mask.append(this_mask)
+        y = np.array(y)
+        X = np.concatenate(X, axis=-1)
+        mask = np.sum(mask, axis=0) > .5 * len(data.zmaps)
+        mask = np.logical_and(mask, np.all(np.isfinite(X), axis=-1))
+        X = X[mask, :].T
+        tmp = np.zeros(list(mask.shape) + [len(X)])
+        tmp[mask, :] = X.T
+        mask_img = nibabel.Nifti1Image(mask.astype(np.int), affine)
+        X = nibabel.four_to_three(nibabel.Nifti1Image(tmp, affine))
+        data = Bunch(zmaps=data.zmaps, X=X, y=y, mask_img=mask_img)
     return data
